I forgot this wasn't a javascript project
Anyways, I got this workign in default tsx format. Use it as a frame of
reference.

DON"T MAKE IT TO A TS: import as the path of the file, don't import the
extension. It won't work. I know from doing it bad:  the  ∑ of the
confusion I was in